requirements.txt
africastalking==2.0.2
aiofiles==25.1.0
aiogram==3.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
argcomplete==3.6.3
arq==0.27.0
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
av==16.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
billiard==4.2.4
build==1.4.0
cbor2==5.8.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
comtypes==1.4.16
constantly==23.10.4
coverage==7.13.4
cron-descriptor==1.4.5
cryptography==46.0.5
ctranslate2==4.7.1
customtkinter==5.2.2
daphne==4.2.1
darkdetect==0.8.0
dataclasses-json==0.6.7
defusedxml==0.7.1
distro==1.9.0
Django==6.0.2
django-celery-beat==2.9.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-extensions==4.1
django-picklefield==3.4.0
django-q2==1.9.0
django-ratelimit==4.1.0
django-redis==6.0.0
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
docstring_parser==0.17.0
docutils==0.22.4
ecdsa==0.19.1
et_xmlfile==2.0.0
faster-whisper==1.2.1
filelock==3.25.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
h11==0.16.0
hf-xet==1.3.2
hiredis==3.3.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.5.0
hyperlink==21.0.0
id==1.6.1
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
jaraco.classes==3.4.0
jaraco.context==6.1.1
jaraco.functools==4.4.0
jiter==0.13.0
Js2Py==0.74
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.7.0
kombu==5.6.2
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.17
langchain-text-splitters==1.1.1
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.12
magic-filter==1.0.12
markdown-it-py==4.0.0
marshmallow==3.26.2
mdurl==0.1.2
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
mypy_extensions==1.1.0
nh3==0.3.3
numpy==2.4.2
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.24.3
openpyxl==3.1.5
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
passlib==1.7.4
paydunya==1.0.7
pillow==12.1.1
pipwin==0.5.2
pipx==1.8.0
platformdirs==4.9.2
playwright==1.58.0
pluggy==1.6.0
plyer==2.1.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==7.34.0
psutil==7.2.2
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydot==4.0.1
pyee==13.0.1
PyGetWindow==0.0.9
Pygments==2.19.2
pyjsparser==2.7.1
PyJWT==2.11.0
PyMsgBox==2.0.1
pyOpenSSL==25.3.0
pyparsing==3.3.2
pyperclip==1.11.0
pypiwin32==223
PyPrind==2.11.3
pyproject_hooks==1.2.0
PyRect==0.2.0
PyScreeze==1.0.1
pySmartDL==1.3.4
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.12.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
python-jose==3.5.0
python3-openid==3.2.0
pyttsx3==2.99
pytweening==1.2.0
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
qrcode==8.2
readme_renderer==44.0
redis==5.3.1
reportlab==4.4.10
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.26.0
rfc3986==2.0.0
rich==14.3.3
# Editable install with no version control (rolax1212==2.0.5)
-e e:\files
rsa==4.9.1
schedule==1.2.2
schema==0.7.8
sentry-sdk==2.54.0
service-identity==24.2.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
social-auth-app-django==5.7.0
social-auth-core==4.8.5
sounddevice==0.5.5
soundfile==0.13.1
soupsieve==2.8.3
SpeechRecognition==3.14.5
SQLAlchemy==2.0.48
sqlparse==0.5.5
srt==3.5.3
sympy==1.14.0
tenacity==9.1.4
tokenizers==0.22.2
tqdm==4.67.3
twine==6.2.0
Twisted==25.5.0
txaio==25.12.2
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.12.0
urllib3==2.6.3
userpath==1.9.2
uuid_utils==0.14.1
vine==5.1.0
vosk==0.3.45
wcwidth==0.6.0
websockets==16.0
whatsapp-api-client-python==0.0.53
whatsapp-chatbot-python==0.9.9
xxhash==3.6.0
yarl==1.23.0
zope.interface==8.2
zstandard==0.25.0
